fix(validation): refresh PR base once per run
Cache only authoritative GitHub PR-base results, reset the cache at each pre-PR invocation, and support non-origin self-tracking branches.


Copilot-Session: 025a12a0-6114-4f48-b25a-876dbd644429
