+++ scripts/validation/checks_common.py
@@ -142,7 +142,13 @@ def _gh_pr_base_ref_name(repo_root: Path, selector: list[str]) -> _PrViewProbe:
 
 def _gh_base_ref_probe(repo_root: Path) -> tuple[str | None, bool, int, str]:
     """Uncached core of the PR-base query: bare lookup, then retry with the
-    upstream head branch if the bare lookup found no PR (issue #4382).
+    upstream head branch (issue #4382) -- but ONLY when the bare lookup
+    explicitly confirmed no PR AND the upstream is not self-tracking. A
+    transient failure (auth, network, rate limit) must not retry: it could
+    look like a false confirmed "no PR" (round 2 review, item 4).
+    Self-tracking (``git push -u origin HEAD``) would repeat the same
+    confirmed query, so :func:`_is_self_tracking_upstream` (issue #2571)
+    gates that too.
 
     Returns ``(base_ref, cacheable, exit_code, stderr)``; ``cacheable`` is
     True for a found PR or confirmed no-PR, False when transient -- see
@@ -151,9 +157,11 @@ def _gh_base_ref_probe(repo_root: Path) -> tuple[str | None, bool, int, str]:
     probe = _gh_pr_base_ref_name(repo_root, [])
     if probe.base_ref:
         return f"origin/{probe.base_ref}", True, probe.exit_code, probe.stderr
+    if not probe.no_pr_confirmed:
+        return None, False, probe.exit_code, probe.stderr
 
     head = _upstream_head_ref_name(repo_root)
-    if not head:
+    if not head or _is_self_tracking_upstream(repo_root):
         return None, probe.no_pr_confirmed, probe.exit_code, probe.stderr
 
     probe2 = _gh_pr_base_ref_name(repo_root, [head])
@@ -197,6 +205,20 @@ _gh_pr_base_cache: dict[tuple[str, str, str], str | None] = {}
 _gh_pr_base_logged_failures: set[tuple[str, str, str]] = set()
 
 
+def _reset_gh_base_cache() -> None:
+    """Clear the gh PR-base cache and its logged-failure set (item 3).
+
+    The cache key, ``(repo_root, branch, HEAD sha)``, is a proxy for "did
+    the checkout change", not "did the remote PR change": two in-process
+    invocations of ``pre_pr.py`` with an unchanged branch/HEAD would
+    otherwise reuse a prior invocation's answer. ``pre_pr.main()`` calls
+    this once at the top, scoping the cache to one invocation while still
+    sharing it across every gate within that run.
+    """
+    _gh_pr_base_cache.clear()
+    _gh_pr_base_logged_failures.clear()
+
+
 def _gh_base_ref(repo_root: Path) -> str | None:
     """Return ``origin/<baseRefName>`` for the open PR, or None.
 
@@ -204,14 +226,11 @@ def _gh_base_ref(repo_root: Path) -> str | None:
     can pass the result to ``git diff`` directly. :func:`_gh_base_ref_probe`
     is the expensive part (``gh pr view``, ~0.42-0.49s/call); this wrapper
     caches ITS result keyed on ``(repo_root, branch, HEAD sha)`` so both
-    resolver functions share one query per branch/HEAD state.
-
-    A transient failure (auth, network, rate limit) is NOT cached --
-    retrying is correct since the condition may clear. Only success or a
-    verified "no open PR" is cached; the first non-authoritative failure
-    per key is logged once, not once per gate. Returns None without caching
-    when gh is absent from PATH or exits non-zero for any other reason.
-    Retries with the upstream head branch when the bare lookup finds no PR
+    resolver functions share one query per branch/HEAD state. A transient
+    failure is NOT cached (retrying is correct once the condition clears);
+    only success or a confirmed "no open PR" is cached, and the first
+    non-authoritative failure per key is logged once, not once per gate.
+    Retries with the upstream head branch per :func:`_gh_base_ref_probe`
     (issue #4382).
 
     A same-named, but NOT behaviorally identical, helper lives in
@@ -248,13 +267,11 @@ def _is_self_tracking_upstream(repo_root: Path) -> bool:
     After ``git push -u origin HEAD``, the branch's upstream is
     ``<remote>/<current_branch>`` -- it tracks itself, emptying a diff
     against it once pushed and silently no-opping "since base" gates
-    (Issue #2571).
-
-    Compares ``branch.<branch>.merge`` (:func:`_upstream_head_ref_name`) to
-    the branch name, instead of string-comparing abbreviated ``@{u}``
-    against a hardcoded ``origin/<branch>``, which missed a non-``origin``
-    fork remote and branch names containing ``/`` (no marked remote/branch
-    boundary in the abbreviated form). ``@{u}`` resolvability is checked
+    (Issue #2571). Compares ``branch.<branch>.merge``
+    (:func:`_upstream_head_ref_name`) to the branch name rather than
+    string-comparing abbreviated ``@{u}`` against a hardcoded
+    ``origin/<branch>``, which missed a non-``origin`` fork remote and
+    branch names containing ``/``. ``@{u}`` resolvability is checked
     first, so a branch with no upstream returns False.
     """
     clean_env = _git_subprocess_env()
@@ -295,24 +312,18 @@ def _is_self_tracking_upstream(repo_root: Path) -> bool:
 def _resolve_branch_base_ref(repo_root: Path) -> str | None:
     """Resolve the branch base ref by trying signals in priority order.
 
-    Not cached itself: a prior version cached the whole resolution via a
-    bare ``functools.cache`` with no invalidation hook, so a branch switch
-    or new commit mid-process kept serving the first answer. The expensive
-    step, the network round trip inside :func:`_gh_base_ref` (``gh pr
-    view``, ~0.42-0.49s/call), now caches its own result keyed on
-    ``(repo_root, branch, HEAD sha)``, invalidating on a branch switch or
-    new commit; the rest here is local, sub-millisecond git plumbing.
+    Not cached itself (a bare ``functools.cache`` on the whole resolution
+    had no invalidation hook); the expensive step, the network round trip
+    inside :func:`_gh_base_ref` (``gh pr view``, ~0.42-0.49s/call), caches
+    its own result keyed on ``(repo_root, branch, HEAD sha)``. The rest
+    here is local, sub-millisecond git plumbing.
 
-    Tries each candidate in order and returns the first one that resolves
-    to a real ref locally: (1) the PR's actual baseRefName via ``gh pr
-    view``, validated with ``git rev-parse --verify`` so an unfetched ref
-    falls through; (2) the current branch's configured upstream via
-    ``@{u}``, EXCEPT when self-tracking (the branch's own
-    ``origin/<branch>`` after ``git push -u origin HEAD`` -- an empty diff
-    against HEAD that would let "since base" gates miss real changes, the
-    bug behind Issue #2571); (3) the remote's default branch via
-    ``refs/remotes/origin/HEAD``; (4) ``origin/main`` as a last resort.
-
+    Priority: (1) the PR's actual baseRefName via ``gh pr view``, validated
+    with ``git rev-parse --verify`` so an unfetched ref falls through; (2)
+    the current branch's ``@{u}``, EXCEPT when self-tracking (``origin/
+    <branch>`` after ``git push -u origin HEAD``, which would empty-diff
+    against HEAD and hide real changes, Issue #2571); (3) the remote
+    default branch via ``refs/remotes/origin/HEAD``; (4) ``origin/main``.
     Returns None when none resolve.
 
     A related helper (``_detect_default_base_ref``) in
@@ -356,22 +367,18 @@ def _resolve_default_base_ref(repo_root: Path) -> str | None:
     """Resolve the base ref for "what changed vs the default branch".
 
     Unlike :func:`_resolve_branch_base_ref`, this deliberately EXCLUDES the
-    current branch's own upstream (``@{u}``). For a feature branch ``@{u}`` is
-    ``origin/<feature-branch>``; once the branch is pushed, ``@{u}...HEAD`` is
-    empty, so a change-vs-base diff misses everything the branch added. That is
-    exactly how ``pre_pr.py`` reported "No changed workflow files" while the
-    pre-push hook (which diffs against the merge-base with ``origin/main``)
-    found and validated one (issue #2571).
+    current branch's own upstream (``@{u}``): for a feature branch ``@{u}``
+    is ``origin/<feature-branch>``, and once pushed, ``@{u}...HEAD`` is
+    empty, so a change-vs-base diff misses everything the branch added
+    (issue #2571; ``pre_pr.py`` reported "No changed workflow files" while
+    the pre-push hook, diffing against the merge-base with ``origin/main``,
+    found one).
 
-    Priority order:
-
-        1. The PR's actual baseRefName via ``gh pr view`` (validated;
-           shares :func:`_gh_base_ref`'s cache with
-           :func:`_resolve_branch_base_ref`).
-        2. The remote default branch via ``refs/remotes/origin/HEAD``.
-        3. ``origin/main`` then local ``main`` as last-resort literals.
-
-    Returns None when none resolve.
+    Priority: (1) the PR's baseRefName via ``gh pr view`` (validated;
+    shares :func:`_gh_base_ref`'s cache with :func:`_resolve_branch_base_ref`);
+    (2) the remote default branch via ``refs/remotes/origin/HEAD``; (3)
+    ``origin/main`` then local ``main`` as last-resort literals. Returns
+    None when none resolve.
     """
     candidates: list[str] = []
     pr_base = _gh_base_ref(repo_root)
@@ -391,18 +398,13 @@ def _resolve_default_base_ref(repo_root: Path) -> str | None:
 def _refresh_remote_base(base_ref: str, repo_root: Path) -> str | None:
     """Best-effort fetch of ``origin/<branch>`` to keep the base ref fresh (#2453).
 
-    The wrapper invokes a base-ref-using validator with the local
-    ``origin/<branch>`` ref. When the local ref is stale (the developer has
-    not fetched in a while), the validator compares against an old version
-    and false-PASSes a bump that is actually insufficient against the real
-    remote. Refreshing here, in the wrapper, keeps the validator itself pure
-    and offline-safe.
+    A stale local ``origin/<branch>`` lets a validator false-PASS a bump
+    that is insufficient against the real remote; refreshing here keeps the
+    validator itself pure and offline-safe.
 
-    Returns:
-        ``None`` when no fetch was attempted (non-``origin/<branch>`` ref, or
-        running under CI where the CI runner already fetched). A short error
-        string when the fetch was attempted and failed; the caller emits a
-        WARNING and proceeds. Empty string on a successful fetch.
+    Returns: None when no fetch was attempted (non-``origin/<branch>`` ref,
+    or under CI where the runner already fetched); a short error string on
+    a failed attempt (caller warns and proceeds); empty string on success.
     """
     if not base_ref.startswith("origin/"):
         return None
@@ -452,14 +454,13 @@ def _run_build_script_gate(
 
     Args:
         repo_root: Repository root path.
-        script_name: Filename under ``build/scripts/`` (e.g.,
+        script_name: Filename under ``build/scripts/`` (e.g.
             ``validate_install_parity.py``).
-        gate_label: Human-readable name for error messages (e.g.,
+        gate_label: Human-readable name for error messages (e.g.
             ``install-parity``).
 
-    Returns:
-        True if the script exits 0, False otherwise. Fails closed when
-        the script is absent or when the base ref cannot be resolved.
+    Returns: True if the script exits 0. Fails closed (False) when the
+        script is absent or the base ref cannot be resolved.
     """
     script = repo_root / "build" / "scripts" / script_name
     if not script.exists():
